Добавлен абстрактный класс
