+++ src/view/abstract.js
@@ -1,18 +1,16 @@
 import {createElement} from "../utils/render";
 
-const createShowMoreButton = () => {
-  return (
-    `<button class="films-list__show-more">Show more</button>`
-  );
-};
-
-export default class LoadMoreButton {
+export default class AbstractComponent {
   constructor() {
+    if (new.target === AbstractComponent) {
+      throw new Error(`Can't instantiate AbstractComponent, only concrete one.`);
+    }
+
     this._element = null;
   }
 
   getTemplate() {
-    return createShowMoreButton();
+    throw new Error(`Abstract method not implemented: getTemplate`);
   }
 
   getElement() {
