+++ src/view/films-list.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import {createElement} from "../utils/render";
 
 const createFilmsListTemplate = () =>
   `<section class="films">
