+++ src/view/popup.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import {createElement} from "../utils/utils.js";
 import {emojies} from "../const";
 
 const createGenresMarkup = (genres) => {
